Implement confirming borrow requests

+++ contracts/Clock.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.4.23;
 
 contract Clock {
-  function getTime() public view returns(uint) {
-    return now;
+  function getTime() public view returns(uint64) {
+    return uint64(now);
   }
 }
